backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
python-dotenv==1.0.0
pydantic==2.5.0
yfinance==0.2.28
python-multipart==0.0.6
httpx==0.27.0
pandas==2.1.4
kiteconnect==4.2.0
nsetools==1.0.6

